+++ src/components/features/media/server-download-modal.tsx
@@ -318,6 +318,7 @@ export function ServerDownloadModal({ entry, open, onOpenChange }: ServerDownloa
                 supportsSmartSearch={torrentStream.selectedProviderSupportsSmartSearch}
                 torrents={torrentStream.torrents}
                 torrentMetadataByInfoHash={torrentStream.torrentMetadataByInfoHash}
+                debridInstantAvailability={torrentStream.debridInstantAvailability}
                 usePreviousBatch={torrentStream.usePreviousBatch}
                 resolution={torrentStream.resolution}
                 searchAcrossProviders={torrentStream.searchAcrossProviders}